+++ pages/my-goals/index.js
@@ -1,0 +1,3 @@
+export default function MyGoalsOPage({}) {
+  return;
+}
